requirements.txt
fastapi
uvicorn
python-multipart
torch
torchaudio
transformers
openai-whisper
mediapipe==0.10.14
ffmpeg-python
noisereduce
edge-tts
scikit-learn
